[HFH-3309] Revert switch to npm local server (#218)
In a recent [renovate bot
PR](https://github.com/powerhome/power-web-development-interview/pull/210/files)
we missed that it swapped the public npm server with our private,
internal, one.

That turns out to be _bad_ as candidates can't pull some of the
libraries.

This reverts that last change to put it back the way it was prior to
that PR.